proc_macro: move some implementation details to a rustc module.
